Added 'cursor' and 'smartcat' to the lists of desired tools on darwin-configuration.nix and default.nix respectively. 'Cursor' was added to 'casks' and 'smartcat' in 'tools/default.nix'. Removed 'Magnet' and 'TickTick' and added 'Xcode' in 'darwin-configuration.nix' -> 'systemPreferences' ->  'masExtraApps'.
Simplified and refactored the 'gcai', 'gbai' and 'prdai' functions in 'modules/zsh/functions.sh' to streamline the commit message, branch name, and PR description generation process respectively with the smartcat tool.

Added a new shell script, 'from_heigh_to_jpg.sh', which converts the '.heic' files in a given directory to '.jpg' and saves them in a new directory, 'jpg_output', within the provided directory. If no directory argument is provided, it defaults to the current directory. The script also checks for the presence of the 'mogrify' command and 'ImageMagick' package before processing.